feat: html 변조 된 값에 대한 예외상황 처리

+++ src/main/java/store/mybooks/front/order/dto/request/BookInfoRequest.java
@@ -25,6 +25,6 @@ public class BookInfoRequest {
     private Integer saleCost;
     private Integer bookCost;
     private Integer selectWrapId;
-    private Integer selectCouponId;
+    private Long selectCouponId;
 
 }
